Refactored pipeline.js and fixed attachments
1. Refactored pipeline.js to be /only/ the pipeline, pulling the rest of
the code, ie steps in thep pipeline, into other files. Goal is that
anything that is to be done to a file/page goes in here, but nothing
else
2. Fixed the uniorg-attach, didn't initially pull that module into this
project, and updated the attachment directory to not be relative. I will
likely need to explicitly =cp= over that =.attach/= directory in the org
build step since hidden directories don't always copy over nicely